requirements.txt
astroid==2.4.2
bcrypt==3.2.0
blinker==1.4
cffi==1.14.2
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
